requirements.txt
flask==3.1.3
aiohttp==3.13.5
requests==2.33.1
python-dotenv==1.2.2
